requirements.txt
python-jose[cryptography]==3.3.0
fastapi==0.111.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
alembic==1.13.1